Added Flashlight Toggle Click Sound and Fixed Bug with Flashlight During Pause Menu
Added a click sound to the spotlight/flashlight game object where if the F key is pressed the audio sound will be played.

Prior to the bug fix, player could toggle on and off the flashlight tool while being in the pause menu.

After the fix, the player cannot toggle the flashlight on or off while in the pause menu.