Storyboard Animation Fix and Added Pause Menu Confirmation Interfaces for MainMenu/Restart and Quit Buttons
Confirmation Interfaces will appear when clicking on main menu button, restart button and quit button in order to check if the user really wants to continue with this action.

Applied fix for storyboard animation not loading up when trying to load the MainGame scene again.

(Fix Can be developed further to mitigate this error in other scenes)