Implemented the Parchment Interaction in Garage Level
A parchment has been added to the level which will provide hints of the level.
When the player is in range of the parchment, they can press E and then a pop up image will show up where the player can read the contents of the parchment consisting of the clues.